refactored file structure and detail fragment
