+++ moVirt/src/main/java/org/ovirt/mobile/movirt/facade/StorageDomainFacade.java
@@ -3,6 +3,7 @@ package org.ovirt.mobile.movirt.facade;
 import android.content.Context;
 import android.content.Intent;
 import android.database.Cursor;
+
 import org.androidannotations.annotations.Bean;
 import org.androidannotations.annotations.EBean;
 import org.ovirt.mobile.movirt.model.EntityMapper;
@@ -11,6 +12,7 @@ import org.ovirt.mobile.movirt.model.trigger.Trigger;
 import org.ovirt.mobile.movirt.rest.OVirtClient;
 import org.ovirt.mobile.movirt.sync.SyncAdapter;
 import org.ovirt.mobile.movirt.ui.storage.StorageDomainDetailActivity_;
+
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
