+++ moVirt/src/main/java/org/ovirt/mobile/movirt/model/DataCenter.java
@@ -6,6 +6,7 @@ import android.net.Uri;
 
 import com.j256.ormlite.field.DatabaseField;
 import com.j256.ormlite.table.DatabaseTable;
+
 import org.ovirt.mobile.movirt.provider.OVirtContract;
 import org.ovirt.mobile.movirt.util.CursorHelper;
 import org.ovirt.mobile.movirt.util.ObjectUtils;
