+++ moVirt/src/main/java/org/ovirt/mobile/movirt/model/trigger/TriggerResolver.java
@@ -1,7 +1,6 @@
 package org.ovirt.mobile.movirt.model.trigger;
 
 import org.ovirt.mobile.movirt.model.BaseEntity;
-import org.ovirt.mobile.movirt.model.Event;
 
 import java.util.Collection;
 import java.util.List;
