+++ moVirt/src/main/java/org/ovirt/mobile/movirt/sync/SyncAdapter.java
@@ -30,9 +30,9 @@ import org.ovirt.mobile.movirt.model.Vm;
 import org.ovirt.mobile.movirt.model.trigger.Trigger;
 import org.ovirt.mobile.movirt.provider.ProviderFacade;
 import org.ovirt.mobile.movirt.rest.OVirtClient;
+import org.ovirt.mobile.movirt.ui.MainActivityFragments;
 import org.ovirt.mobile.movirt.ui.MainActivity_;
 import org.ovirt.mobile.movirt.util.NotificationHelper;
-import org.ovirt.mobile.movirt.ui.MainActivityFragments;
 
 import java.util.ArrayList;
 import java.util.Collection;
