+++ moVirt/src/main/java/org/ovirt/mobile/movirt/ui/LoaderFragment.java
@@ -1,6 +1,7 @@
 package org.ovirt.mobile.movirt.ui;
 
 import android.support.v4.app.Fragment;
+
 import org.androidannotations.annotations.EFragment;
 
 @EFragment
