+++ moVirt/src/main/java/org/ovirt/mobile/movirt/ui/dashboard/DashboardActivity.java
@@ -2,6 +2,7 @@ package org.ovirt.mobile.movirt.ui.dashboard;
 
 import android.view.WindowManager;
 import android.widget.ProgressBar;
+
 import org.androidannotations.annotations.AfterViews;
 import org.androidannotations.annotations.EActivity;
 import org.androidannotations.annotations.ViewById;
