+++ moVirt/src/main/java/org/ovirt/mobile/movirt/ui/dashboard/DashboardGeneralFragment.java
@@ -6,6 +6,7 @@ import android.support.v4.app.LoaderManager;
 import android.support.v4.content.Loader;
 import android.util.Log;
 import android.widget.TextView;
+
 import org.androidannotations.annotations.AfterViews;
 import org.androidannotations.annotations.Bean;
 import org.androidannotations.annotations.EFragment;
@@ -17,6 +18,7 @@ import org.ovirt.mobile.movirt.model.Host;
 import org.ovirt.mobile.movirt.model.StorageDomain;
 import org.ovirt.mobile.movirt.provider.ProviderFacade;
 import org.ovirt.mobile.movirt.ui.LoaderFragment;
+
 import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.List;
