+++ moVirt/src/main/java/org/ovirt/mobile/movirt/ui/storage/StorageDomainFragment.java
@@ -6,6 +6,7 @@ import android.widget.CursorAdapter;
 import android.widget.ImageView;
 import android.widget.SimpleCursorAdapter;
 import android.widget.TextView;
+
 import org.androidannotations.annotations.EFragment;
 import org.ovirt.mobile.movirt.R;
 import org.ovirt.mobile.movirt.model.StorageDomain;
