+++ moVirt/src/main/res/layout/fragment_disk_detail.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.ovirt.mobile.movirt.ui.DiskDetailFragment">
+    tools:context="org.ovirt.mobile.movirt.ui.vms.VmDiskDetailFragment">
 
     <android.support.v4.widget.SwipeRefreshLayout
         android:id="@+id/swipeDisksContainer"
