+++ moVirt/src/main/res/layout/fragment_event_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.ovirt.mobile.movirt.ui.EventsFragment">
+    tools:context="org.ovirt.mobile.movirt.ui.events.EventsFragment">
 
     <android.support.v4.widget.SwipeRefreshLayout
         android:id="@+id/swipeEventsContainer"
