+++ moVirt/src/main/res/layout/fragment_nic_detail.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.ovirt.mobile.movirt.ui.NicDetailFragment">
+    tools:context="org.ovirt.mobile.movirt.ui.vms.VmNicDetailFragment">
 
     <android.support.v4.widget.SwipeRefreshLayout
         android:id="@+id/swipeNicsContainer"
